Update the NotFound view to support "relative" redirects within the provided version
This is needed as I broke all previous redirects in the reorganisation, and need to make sure you can get from
    /2.x/en/developing-in-modx/advanced-development/custom-manager-pages/actions-and-menus
to
    /2.x/en/extending-modx/menus/actions
from any old links, favs, indexed URLs etc.

The way this now works is that if it detects a version in the URL (in the above example, 2.x), it takes it out of the URI it looks for. This allows the redirects.json to contain an entry for "en/developing-in-modx/.../actions-and-menus" and still redirect appropriately, without needing to triple the size of redirects.json.

Also add json to the composer.json because phpstorm kept nagging.